==> 00_scripts_raw_data/00_data_preprocessing/05_balance_dataset_oversampling-SMOTE.ipynb <==
# https://www.sharpsightlabs.com/blog/scikit-train_test_split/
#sudo pip install imbalanced-learn
#https://machinelearningmastery.com/random-oversampling-and-undersampling-for-imbalanced-classification/

# example of random oversampling to balance the class distribution
from collections import Counter
from sklearn.datasets import make_classification
from imblearn.over_sampling import RandomOverSampler, SMOTE
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np

# general data


inputFolders = "../data_2019_processed/"
outputFolders = ""

# Dataframe of all datasets
# initialize list of lists
datasets = ['0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs',
            '0tdmm6rwW3KquQ73ATYYJ5JkpMtvbppJ0VzA2GExdA', 
            '2cyV53lVyUtlMj0BRwilEWtYJwUiviYoL48cZBPBq0', 
            '2J22RukYnEbKTk7t+iUVDBkorcyL5NKN6TrLe89ys', 
            #'5FLZBTVAPwdq9QezHE2sVCJIs7p+r6mCemA2gp9jATk', #does not have the file
            '7EYF5I04EVqisUJCVNHlqn77UAuOmwL2Dahxd3cA', 
            'a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4', 
            'ae4JJBZDycEcY8McJF+3BxyvZ1619y03BNdCxzpZTc', 
            'Ch3u5Oaz96VSrQbf0z31X6jEIbeIekkC0mwPzCdeJ1U', 
            'CH8f0yZkZL13zWuE9ks1CkVJRVrr+jsGdUXHrZ6YeA', 
            'DHO1K4jgiwZJOfQTrxvKE2vn7hkjamigroGD5IaeRc', 
            #'DHPqzSqSttiba1L3BD1cptNJPjSxZ8rXxF9mY3za6WA', # does not have asleep data
            'dQEFscjqnIlug8Tgq97JohhSQPG2DEOWJqS86wCrcY', 
            'HFvs2CohmhHte+AaCzFasjzegGzxZKPhkrX23iI6Xo', 
            'jgB9E8v3Z6PKdTRTCMAijBllA9YEMtrmHbe4qsbmJWw', 
            'JkY++R7E8myldLN3on6iQ78Ee78zCbrLuggfwGju3I', 
            'K4SLohf+TN1Ak8Dn8iE3Lme7rEMPISfppB2sXfHX8', 
            'oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q', 
            'pyt24oiDAHsmgWMvkFKz2fn2pwcHiXchd6KchLM', 
            #'PZCf1nfvhR+6fk+7+sPNMYOgb8BAMmtQtfoRS83Suc', # does not have asleep data
            'QUNCATForxzK0HHw46LrGOMWh0eVA8Y5XWEiUXX+cQ', 
            'rIl2UK9+bQ+tzpFdbJAdbBxEa5GbgrgC030yEaENLw', 
            'RoBW3cDOO9wWRMPO2twQff83MPc+OXn6gJ+a1DafreI', 
            'SH3kQeyd5volraxw8vOyhlowNqWBPr1IJ9URNXUL4', 
            'VVpwFNMrEglveh6MDN8lrRzTy5OwzglD4FURfM4A2is', 
            'Wa1mcNmbh66S7VS6GIzyfCFMD3SGhbtDQyFP1ywJEsw', 
            'XCKRE0BWRHxfP1kZIihgtT+jUjSp2GE8v5ZlhcIhVmA', 
            'YI5Y79K6GXqAUoGP6PNyII8WKlAoel4urDxWSVVOvBw', 
            'ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM', 
            'ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY']

# Create the pandas DataFrame
folders = pd.DataFrame(datasets, columns=['folder'])

#folders

pd.read_csv(inputFolders+"student_0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs_transformed.csv")

# load cliend data
clientList = []

# balancing classes
#oversample = RandomOverSampler(sampling_strategy='auto') #minority
smoteOversampling = SMOTE(sampling_strategy='auto',random_state=32)
#undersample = RandomUnderSampler(sampling_strategy='majority')

for folder in datasets:
    print (" ", str(inputFolders)+"student_"+folder+"_transformed.csv") #_numeric
    # load client data
    dataset = pd.read_csv(inputFolders+"student_"+folder+"_transformed.csv")
    
    # revome text date
    
    if 'timestamp_text' in dataset.columns:
        del  dataset['timestamp_text']
    
    dataset['class'] = dataset['class'].replace(['awake'], '0')
    dataset['class'] = dataset['class'].replace(['asleep'], '1')
    
    # print(dataset)
    y_train = dataset['class'].copy()
    print("initial len: ", dataset.shape)

    # does not add datasets that dont have instances from both classes
    if y_train.sum() != 0 and (y_train.sum() != len(y_train)):
        
        # do balancing technique
        X_train2, y_train2 = smoteOversampling.fit_resample(dataset, y_train)
        #X_train_under, y_train_under = undersample.fit_resample(X_train_under, y_train_under)
        
        dataset['class'] = y_train
        X_train2['class'] = y_train2
        
        clientList.append(dataset)
        
        dataset['class'].astype(int) 
        dataset['class'].astype(str) 
        dataset['class'] = dataset['class'].replace('0','awake')
        dataset['class'] = dataset['class'].replace('1','asleep' )
    
        # summarize data shape
        print("final len: ", X_train2.shape)
        print("----")
        # summarize class distribution
        print("Before oversampling: ", Counter(y_train))
        print("After undersampling: ", Counter(y_train2))
        dataset.to_csv(inputFolders+"student_"+folder+"_transformed_smote.csv", sep=',', encoding='utf-8', index=False)
        print("----")
        print("----")
        print("----")
        
print("Total",(len(clientList)))

pd.read_csv(inputFolders+"student_ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY_transformed_smote.csv")

dataset
